+++ apps/kg/kvclient.py
@@ -10,6 +10,7 @@ if os.name != 'nt':
 
 import torch.multiprocessing as mp
 from train_pytorch import load_model, dist_train_test
+from utils import get_compatible_batch_size
 
 from train import get_logger
 from dataloader import TrainDataset, NewBidirectionalOneShotIterator
@@ -164,14 +165,6 @@ def get_local_machine_id(server_namebook):
             break
 
     return res
-
-def get_compatible_batch_size(batch_size, neg_sample_size):
-    if neg_sample_size < batch_size and batch_size % neg_sample_size != 0:
-        old_batch_size = batch_size
-        batch_size = int(math.ceil(batch_size / neg_sample_size) * neg_sample_size)
-        print('batch size ({}) is incompatible to the negative sample size ({}). Change the batch size to {}'.format(
-            old_batch_size, neg_sample_size, batch_size))
-    return batch_size
 
 
 def start_worker(args, logger):
